+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - Inspect View: Truncate display of large sample summary fields to improve performance.
+- Inspect View: Improve reliability of 'View Log' link handling in VS Code.
 - Bugfix: Truncate large target fields in sample summaries.
 
 ## 0.3.148 (21 November 2025)